"fixed" the bugs related to the turn system
The player now has a set amount of moves in a turn. The bugs related to the negative values should be fixed.

The way the fix works is not preferable but it does the job (aka I can't find the one line that makes it bug so I added another one to hide my mistake #relatable)